Move platform-specific URL interface into main engine.
On Android and iOS, transmitting data to/from URLs is performed using
a platform-specific API, rather than revliburl + raw sockets.  The
implementation was previously in mobile-specific source files.

Emscripten also requires the use of platform-specific APIs for
communication over HTTP, but is treated as a desktop platform.

This patch moves the platform-specific implementation of
`MCS_geturl()` etc. into a source file that is compiled on all
platforms but where the implementations are compiled only if the
`FEATURE_PLATFORM_URL` feature macro is defined.
